realise first page on warehouses global changed

+++ app/src/main/java/ru/wb/go/db/entity/courier/CourierWarehouseLocalEntity.kt
@@ -13,5 +13,5 @@ data class CourierWarehouseLocalEntity(
     val fullAddress: String,
     val longitude: Double,
     val latitude: Double,
-    val distanceFromUser:Float
+    val distanceFromUser:Int
 )
